basicwidget: update a text value immediately when it is not editable
For example, a dropdown list item's text size is not measured correctly
until it is shown. This caused a rendering shift issue in examples/gallery.
This change fixed this issue.

The purpose is the same as d4aa97ca415975cccf545b5efe5a7ddb57bdc95f,
but the approach is different.